// src/app/interfaces/customer.ts
export interface Customer {
    key?: string;
    first_name?: string;
    last_name?: string;
    nationality?:string;
    title?: string;
    phone_number?: string;
    email?: string;
    address?: string;
    city?: string;
    country?: string;
    university?: string;
    zip_code?: string;
    emergency_contact?: string;
    emc_phone_number?: string;
    emc_email?: string;
    status?: number;
    createdate?: Date;
    createby?: object;
    updatedate?: Date;
    updateby?: object;
}